Debug xcodebuild execution
It has been reported that the `xcodebuild` process spawned from the
`push` command can block indefinitely.

This branch is an on-going investigation of the issue to find the real
culprit.

* Extra debug log is added before and after the call to the `run()`
method of the `xcodebuild` process so that we can identify whether it's
the `run()` call that blocks indefinitely or the reading file handles.
* On MacOS 10.15.4+ systems the `readToEnd()` method is used instead of
the deprecated `readDataToEndOfFile()`. Those methods are synchronous
and block until the whole data is read. (That's currently what is
believed to be causing the deadlock).